requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pg8000==1.31.2
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
cloudinary==1.40.0
qrcode[pil]==7.4.2
pillow==10.3.0
python-multipart==0.0.9
pydantic[email]==2.7.1
aiofiles==25.1.0